web-app/requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# OpenAI Integration
openai>=1.50.0

# Vector Database and Embeddings
chromadb==0.4.15
sentence-transformers>=3.0.0

# Environment Variables
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

